+++ config/src/main/java/com/alibaba/nacos/config/server/auth/ConfigResourceParser.java
@@ -16,11 +16,11 @@
 
 package com.alibaba.nacos.config.server.auth;
 
-import com.alibaba.nacos.core.auth.Resource;
-import com.alibaba.nacos.core.auth.ResourceParser;
 
 import javax.servlet.http.HttpServletRequest;
 
+import com.alibaba.nacos.auth.model.Resource;
+import com.alibaba.nacos.auth.parser.ResourceParser;
 import org.apache.commons.lang3.StringUtils;
 
 /**
