+++ console/src/main/java/com/alibaba/nacos/console/exception/ConsoleExceptionHandler.java
@@ -16,8 +16,8 @@
 
 package com.alibaba.nacos.console.exception;
 
+import com.alibaba.nacos.auth.exception.AccessException;
 import com.alibaba.nacos.common.utils.ExceptionUtil;
-import com.alibaba.nacos.core.auth.AccessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
