+++ console/src/main/java/com/alibaba/nacos/console/security/nacos/JwtTokenManager.java
@@ -16,7 +16,7 @@
 
 package com.alibaba.nacos.console.security.nacos;
 
-import com.alibaba.nacos.core.auth.AuthConfigs;
+import com.alibaba.nacos.auth.common.AuthConfigs;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
